+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
 - Allow customizing the storage disk to use [#6](https://github.com/swisnl/openapi-spec-generator/pull/6).
 - Add support for `Has`, `WhereNull` and `WhereNotNull` filters.
 
+### Fixed
+- Use correct description for `WherePivotNotIn` filter.
+
 
 ## [0.4.0] - 2022-02-24